requirements.txt
annotated-types==0.7.0
anyio==4.11.0
blockbuster==1.5.25
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cloudpickle==3.1.1
cryptography==44.0.3
distro==1.9.0
easyocr==1.7.2
fastapi==0.115.13
forbiddenfruit==0.1.4
googleapis-common-protos==1.71.0
grpcio==1.76.0
grpcio-tools==1.76.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.0
jiter==0.11.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema_rs==0.29.1
langchain==1.0.2
langchain-core==1.0.0
langchain-openai==1.0.1
langgraph==1.0.1
langgraph-api==0.4.46
langgraph-checkpoint==3.0.0
langgraph-cli==0.4.4
langgraph-prebuilt==1.0.1
langgraph-runtime-inmem==0.14.1
langgraph-sdk==0.2.9
openai==2.6.0
orjson==3.11.3
ormsgpack==1.11.0
packaging==25.0
Pillow==11.1.0
protobuf==6.33.0
pycparser==2.23
PyPDF2==3.0.1
pydantic==2.12.3
pydantic_core==2.41.4
PyJWT==2.10.1
python-dotenv==1.1.1
python-multipart==0.0.12
PyYAML==6.0.3
regex==2025.10.23
requests==2.32.5
requests-toolbelt==1.0.0
setuptools==80.9.0
sniffio==1.3.1
sse-starlette==2.1.3
starlette==0.46.0
structlog==25.4.0
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
truststore==0.10.4
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
xxhash==3.6.0
zipp==3.23.0
zstandard==0.25.0
motor==3.6.0
